fix(dashboard): restore asynqmon iframe for redis queue monitoring
Route queue monitoring by broker: postgres keeps the native inspector UI;
redis keeps the existing asynqmon embed unchanged.
