+++ web/ui/dashboard/src/app/private/pages/queue-monitoring/queue-monitoring-asynqmon.component.html
@@ -1,0 +1,46 @@
+@if (!hasAsynqLicense()) {
+  <div class="p-20px border border-warning-100 bg-warning-a3 rounded-12px mb-16px">
+    <p class="font-body font-normal text-[14px] leading-normal text-new.text-primary">Your license does not include Asynq monitoring.</p>
+  </div>
+}
+
+@if (hasAsynqLicense()) {
+  @if (sessionStatus === 'error') {
+    <div class="font-body font-normal text-[13px] leading-normal text-danger-11 bg-danger-a3 border border-danger-100 rounded-8px p-12px mb-16px">
+      {{ sessionError }}
+    </div>
+  }
+  <div class="qm-layout" [class.qm-layout--fullscreen]="iframeFullscreen && iframeVisible">
+    <div class="qm-card bg-white-100 border border-new.border rounded-12px p-0 overflow-hidden flex flex-col">
+      <div class="qm-main">
+        @if (sessionStatus === 'minting' && !iframeVisible) {
+          <div class="qm-loading">
+            <convoy-loader position="relative"></convoy-loader>
+          </div>
+        }
+        @if (iframeVisible) {
+          <iframe
+            #monitorIframe
+            title="Asynqmon queue monitoring"
+            class="qm-iframe w-full bg-neutral-2"
+          ></iframe>
+        }
+      </div>
+      <div class="qm-footer flex flex-wrap items-center gap-8px px-16px py-12px border-t border-new.border bg-new.surface-subtle">
+        <span class="font-heading font-medium text-[14px] leading-[20px] text-new.text-primary mr-auto">Queue monitoring</span>
+        <button
+          type="button"
+          class="px-12px py-8px rounded-full bg-white-100 border border-new.border font-body font-medium text-[13px] leading-[16px] text-new.text-primary transition-colors hover:bg-new.surface-muted disabled:opacity-50 disabled:pointer-events-none"
+          [disabled]="!iframeVisible"
+          (click)="toggleIframeFullscreen()"
+          >
+          {{ iframeFullscreen ? 'Exit Fullscreen' : 'Expand Fullscreen' }}
+        </button>
+        <button type="button" class="px-12px py-8px rounded-full bg-new.primary-400 font-body font-medium text-[13px] leading-[16px] text-white-100 transition-all hover:opacity-90 disabled:opacity-50 disabled:pointer-events-none" [disabled]="sessionStatus === 'minting'" (click)="mintSessionAndLoad()">
+          {{ sessionStatus === 'minting' ? 'Refreshing…' : sessionStatus === 'ready' ? 'Refresh' : 'Retry' }}
+        </button>
+      </div>
+    </div>
+    <p class="font-body font-normal text-[12px] leading-normal text-new.text-secondary mt-12px">Asynq queues and tasks</p>
+  </div>
+}
