+++ web/ui/dashboard/src/app/private/pages/queue-monitoring/queue-monitoring-asynqmon.component.ts
@@ -1,0 +1,157 @@
+import { Component, ElementRef, HostListener, OnInit, ViewChild } from '@angular/core';
+
+import { Router } from '@angular/router';
+import axios from 'axios';
+import { apiOrigin } from 'src/app/services/api-origin';
+import { HttpService } from 'src/app/services/http/http.service';
+import { LicensesService } from 'src/app/services/licenses/licenses.service';
+import { RbacService } from 'src/app/services/rbac/rbac.service';
+import { LoaderModule } from 'src/app/private/components/loader/loader.module';
+
+@Component({
+	selector: 'convoy-queue-monitoring-asynqmon',
+	imports: [LoaderModule],
+	templateUrl: './queue-monitoring-asynqmon.component.html',
+	styleUrls: ['./queue-monitoring-asynqmon.component.scss']
+})
+export class QueueMonitoringAsynqmonComponent implements OnInit {
+	private static readonly fullscreenStorageKey = 'CONVOY_QUEUE_MONITORING_FULLSCREEN';
+
+
+	/** Iframe loads here (session cookie path matches this prefix only). */
+	embedMonitoringUrl = '';
+
+	sessionStatus: 'idle' | 'minting' | 'ready' | 'error' = 'idle';
+	sessionError: string | null = null;
+	iframeVisible = false;
+	iframeFullscreen = false;
+
+	@ViewChild('monitorIframe') monitorIframe?: ElementRef<HTMLIFrameElement>;
+
+	constructor(
+		private readonly httpService: HttpService,
+		private readonly rbacService: RbacService,
+		private readonly router: Router,
+		public readonly licenses: LicensesService
+	) {}
+
+	async ngOnInit(): Promise<void> {
+		const role = await this.rbacService.getUserRole();
+		if (role !== 'INSTANCE_ADMIN') {
+			this.router.navigateByUrl('/');
+			return;
+		}
+
+		this.embedMonitoringUrl = this.buildEmbedMonitoringUrl();
+
+		this.loadFullscreenPreference();
+
+		await this.licenses.loadAllLicenses();
+
+		if (this.hasAsynqLicense()) {
+			await this.mintSessionAndLoad();
+		}
+	}
+
+	hasAsynqLicense(): boolean {
+		// Instance license carries asynq_monitoring for self-hosted Premium/Enterprise.
+		// Refreshed after trial start via loadAllLicenses() above.
+		return this.licenses.hasInstanceLicense('AsynqMonitoring');
+	}
+
+	@HostListener('document:keydown.escape')
+	onEscape(): void {
+		if (this.iframeFullscreen) {
+			this.iframeFullscreen = false;
+			this.persistFullscreenPreference();
+		}
+	}
+
+	toggleIframeFullscreen(): void {
+		this.iframeFullscreen = !this.iframeFullscreen;
+		this.persistFullscreenPreference();
+	}
+
+	async mintSessionAndLoad(): Promise<void> {
+		this.sessionStatus = 'minting';
+		this.sessionError = null;
+		this.iframeVisible = false;
+
+		const token = this.getSessionToken();
+		if (!token) {
+			this.sessionStatus = 'error';
+			this.sessionError = 'No dashboard user token found (log in again).';
+			return;
+		}
+
+		try {
+			const sessionUrl = this.buildSessionUrl();
+			await axios.post(sessionUrl, null, {
+				headers: {
+					Authorization: `Bearer ${token}`,
+					'X-Convoy-Version': '2024-04-01'
+				},
+				withCredentials: true
+			});
+
+			this.sessionStatus = 'ready';
+			this.iframeVisible = true;
+
+			setTimeout(() => {
+				if (this.monitorIframe?.nativeElement) {
+					this.monitorIframe.nativeElement.src = this.buildEmbedReloadUrl();
+				}
+			}, 100);
+		} catch (e: unknown) {
+			this.sessionStatus = 'error';
+			this.sessionError = e instanceof Error ? e.message : String(e);
+		}
+	}
+
+	private getSessionToken(): string | null {
+		// Queue monitoring session minting requires the dashboard user token.
+		const userToken = this.httpService.authDetails()?.access_token || null;
+		if (!userToken) {
+			return null;
+		}
+
+		return this.normalizeToken(userToken);
+	}
+
+	private normalizeToken(token: string): string {
+		return token.replace(/^Bearer\s+/i, '').trim();
+	}
+
+	private apiBase(): string {
+		return apiOrigin();
+	}
+
+	private buildEmbedMonitoringUrl(): string {
+		return `${this.apiBase()}/queue/monitoring/embed/`;
+	}
+
+	private buildEmbedReloadUrl(): string {
+		const separator = this.embedMonitoringUrl.includes('?') ? '&' : '?';
+		return `${this.embedMonitoringUrl}${separator}_r=${Date.now()}`;
+	}
+
+	private buildSessionUrl(): string {
+		return `${this.apiBase()}/queue/monitoring/session`;
+	}
+
+	private loadFullscreenPreference(): void {
+		try {
+			this.iframeFullscreen = localStorage.getItem(QueueMonitoringAsynqmonComponent.fullscreenStorageKey) === 'true';
+		} catch {
+			this.iframeFullscreen = false;
+		}
+	}
+
+	private persistFullscreenPreference(): void {
+		try {
+			localStorage.setItem(QueueMonitoringAsynqmonComponent.fullscreenStorageKey, this.iframeFullscreen ? 'true' : 'false');
+		} catch {
+			/* private mode / quota */
+		}
+	}
+}
